Fix GaussDB debugger issues found by SWTBot acceptance

+++ plugins/org.jkiss.dbeaver.ext.gaussdb.debug.core/src/org/jkiss/dbeaver/ext/gaussdb/debug/core/internal/GaussDBDebugSessionInfo.java
@@ -9,7 +9,7 @@ import org.jkiss.dbeaver.debug.DBGSessionInfo;
 
 import java.util.Map;
 
-public record GaussDBDebugSessionInfo(int processId, String nodeName, int port) implements DBGSessionInfo {
+public record GaussDBDebugSessionInfo(long processId, String nodeName, int port) implements DBGSessionInfo {
     @Override
     public Object getID() {
         return processId;
